pyproject.toml
[tool.poetry]
name = "lastfmcollagegenerator-web"
version = "0.1.0"
description = ""
authors = ["Pau Riera <priera@bmat.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
jinja2 = "^3.1.2"
lastfmcollagegenerator = "^0.3.0"
aiofiles = "^0.8.0"
python-dotenv = "^0.20.0"
python-multipart = "^0.0.5"
jinja2-simple-tags = "^0.4.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
